Backend/nowic-backend/requirements.txt
django==5.0.6
django-ninja==1.3.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.1
pyjwt[crypto]==2.8.0
svix==1.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
requests==2.31.0
whitenoise==6.6.0
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
Pillow==10.3.0
sentry-sdk[django]==1.45.0
pytest==8.2.0
pytest-django==4.8.0
pytest-asyncio==0.23.7
pytest-mock==3.14.0
pytest-env==1.1.3
factory-boy==3.3.0
faker==25.2.0
black==24.4.2
flake8==7.0.0
isort==5.13.2
celery==5.3.6
django-celery-beat==2.6.0
redis==5.0.4